ecommerce-chatbot/backend/requirements.txt
django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
psycopg2-binary==2.9.10
faker==37.3.0
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.2.0 